refactor[core]: define uploader subpackage
