+++ pixl_core/src/core/uploader/__init__.py
@@ -1,0 +1,22 @@
+#  Copyright (c) University College London Hospitals NHS Foundation Trust
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+"""
+Uploader package
+Contains base uploader class and
+- FTPS uploader class
+- **in progress** Azure uploader class
+- **in progress** DICOMWeb uploader class
+
+Uploader class gets appropriate secret credentials from Azure key vault and uploads data
+"""
