+++ pytest-pixl/tests/samples_for_fixture_tests/test_ftpserver_fixture/test_ftpserver_login.py
@@ -15,7 +15,7 @@
 
 from pathlib import Path
 
-from core.upload import _connect_to_ftp
+from core.uploader.ftps import _connect_to_ftp
 
 TEST_FILE_CONTENT = "test text"
 TEST_FILENAME = "testfile.txt"
